Crypto feature - Add revolut export handling (#15)
* add the revolut csv line, a mapper to crypto transaction will not be enough because for exchange crypto to crypto we have two lines exchange one for each currenvy

* update the loading page to be consistent between screens

* add some margin on the loader inside the calculte button and add a 2mb size max for the file input

* finalize the tdd oriented development, still need to do an actual test with my revolut export file

* remove unused properties from the revolut transaction record

* finalize the revolut crypto handling but it does not match the result from the waltio export, to be tested a bit more

* add an order bvyt on the retraits in crypto service

* remove some properties with no business value from the retrait model and add some tests to start on the revolut process coverage, it demands some refactoring on the crypto domain because we are loosing some precision with double instead of decimal type

* update all double as decimal and add a rounding on the gains in euros report field

* analyze the issue was ordering in the revolut transactions, it was not sorted, add unit test and fix the issue, add end to end test for revolut

* revert useless tolist in revolut service

* simplify exception

* remove the public attribute in the interface

+++ RevoProfit.Core/Crypto/Models/CryptoReport.cs
@@ -3,6 +3,6 @@
 public class CryptoReport
 {
     public int Year { get; init; }
-    public double GainsEnEuros { get; init; }
-    public double FraisEnEuros { get; init; }
+    public decimal GainsEnEuros { get; init; }
+    public decimal FraisEnEuros { get; init; }
 }
